+++ src/main/java/net/countercraft/movecraft/warfare/features/Warfare.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.warfare.utils;
+package net.countercraft.movecraft.warfare.features;
 
 // Generic object which is the base for Sieges and Assaults
 public class Warfare {
